new branch with staging for streaming data generator and rphashobject , also changed byte[] decode to long[] , spherical is a multi-test decoder, where if any of the 4 decodings matches, the vectors are considered close.

+++ src/edu/uc/rphash/decoders/Decoder.java
@@ -2,7 +2,7 @@ package edu.uc.rphash.decoders;
 
 public interface Decoder {
 	abstract int getDimensionality();
-	abstract byte[] decode(float[] f);
+	abstract long[] decode(float[] f);
 	abstract float getErrorRadius();
 	abstract float getDistance();
 }
